requirements.txt
# Core Flask dependencies
Flask==3.1.2
Flask-CORS==6.0.1

# HTTP requests
requests==2.32.5

# Environment configuration
python-dotenv==1.1.1

# Development dependencies (optional)
# pytest==8.3.3
# pytest-flask==1.3.0
# flake8==7.1.1
# black==24.8.0